Added generateScreenshots script
Added ScreenshotsGenerator which is an executable Swift Package to generate screenshots on multiple simulators that finally generates an html export of all the screenshots taken from all screens of the app.
Added generateScreenshots shell script that builds the executable from the Swift Package and runs it.

You can run this script like this:
`./generateScreenshots d m:14ProMax,a:dark m:14ProMax m:8`
or
`./generateScreenshots i`